% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraAlertDialog.R
\name{chakraAlertDialog}
\alias{chakraAlertDialog}
\title{Alert dialog widget}
\usage{
chakraAlertDialog(
  inputId,
  options = chakraAlertDialogOptions(),
  openButton,
  header,
  body,
  footerButtons = Tag$Button("Cancel", action = "cancel", value = "cancel")
)
}
\arguments{
\item{inputId}{widget id}

\item{options}{named list of options created with
\code{\link{chakraAlertDialogOptions}}}

\item{openButton}{a chakra button to open the alert dialog}

\item{header}{an \code{AlertDialogHeader} element}

\item{body}{an \code{AlertDialogBody} element}

\item{footerButtons}{a chakra button or a list of chakra buttons}
}
\value{
A widget that can be used in \code{\link{chakraComponent}}.
}
\description{
An alert dialog widget.
}
\examples{
library(shiny)
library(shinyChakraUI)

ui <- chakraPage(

  br(),

  chakraComponent(
    "mycomponent",

    chakraAlertDialog(
      inputId = "alertDialog",
      openButton = Tag$Button(
        leftIcon = Tag$DeleteIcon(),
        colorScheme = "red",
        "Delete customer"
      ),
      header = Tag$AlertDialogHeader(
        fontSize = "lg",
        fontWeight = "bold",
        "Delete customer?"
      ),
      body = Tag$AlertDialogBody(
        "Are you sure? You can't undo this action afterwards."
      ),
      footerButtons = list(
        Tag$Button(
          action = "cancel",
          value = "CANCEL",
          "Cancel"
        ),
        Tag$Button(
          action = "disable",
          value = "DISABLE",
          colorScheme = "red",
          ml = 3,
          "Disable"
        ),
        Tag$Button(
          action = "remove",
          value = "REMOVE",
          ml = 3,
          "Remove"
        )
      )
    )

  )

)

server <- function(input, output, session){

  observe({
    print(input[["alertDialog"]])
  })

}

if(interactive()){
  shinyApp(ui, server)
}
}
