% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraAlertDialogInput}
\alias{chakraAlertDialogInput}
\title{Title}
\usage{
chakraAlertDialogInput(
  inputId,
  options = chakraAlertDialogOptions(),
  openButton,
  header,
  body,
  footerButtons = chakraButton("Cancel", action = "cancel", value = "cancel")
)
}
\arguments{
\item{inputId}{}
}
\value{

}
\description{
Title
}
