% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraAlertDialog.R
\name{chakraAlertDialogOptions}
\alias{chakraAlertDialogOptions}
\title{Alert dialog options}
\usage{
chakraAlertDialogOptions(
  closeOnEsc = TRUE,
  colorScheme = "red",
  isCentered = TRUE,
  motionPreset = "scale",
  size = "md",
  ...
)
}
\arguments{
\item{closeOnEsc}{whether to close the modal on pressing the 'esc' key}

\item{colorScheme}{a chakra color scheme}

\item{isCentered}{whether to center the modal on screen}

\item{motionPreset}{transition that should be used for the modal; one of
\code{"scale"}, \code{"none"}, \code{"slideInBottom"}, or
\code{"slideInRight"}}

\item{size}{modal size, \code{"sm"}, \code{"md"}, \code{"lg"}, \code{"xl"},
\code{"2xl"}, \code{"full"}, \code{"xs"}, \code{"3xl"}, \code{"4xl"},
\code{"5xl"}, or \code{"6xl"}}

\item{...}{other attributes}
}
\value{
A named list, for usage in \code{\link{chakraAlertDialog}}.
}
\description{
Options for the alert dialog widget
  (\code{\link{chakraAlertDialog}}).
}
