% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraBox}
\alias{chakraBox}
\title{Title}
\usage{
chakraBox(
  text,
  as = NULL,
  onHover = NULL,
  onActive = NULL,
  onFocus = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
