% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraButton}
\alias{chakraButton}
\title{Title}
\usage{
chakraButton(
  text,
  value = text,
  action = "none",
  colorScheme = "gray",
  isFullWidth = FALSE,
  leftIcon = NULL,
  rightIcon = NULL,
  size = "md",
  variant = "solid",
  onHover = NULL,
  onActive = NULL,
  onFocus = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
