% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraColorSchemes}
\alias{chakraColorSchemes}
\title{Chakra color schemes}
\usage{
chakraColorSchemes()
}
\description{
List of Chakra color schemes (to use as a \code{colorScheme}
  attribute in e.g. Chakra buttons).
}
\examples{
chakraColorSchemes()
}
