% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraDrawer}
\alias{chakraDrawer}
\title{Title}
\usage{
chakraDrawer(
  inputId,
  openButton,
  options = chakraDrawerOptions(),
  closeButton = TRUE,
  header,
  body,
  footer
)
}
\arguments{
\item{footer}{}
}
\value{

}
\description{
Title
}
