% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraDrawerOptions}
\alias{chakraDrawerOptions}
\title{Title}
\usage{
chakraDrawerOptions(
  closeOnEsc = TRUE,
  closeOnOverlayClick = TRUE,
  colorScheme = NULL,
  id = NULL,
  isCentered = FALSE,
  isFullHeight = FALSE,
  motionPreset = "scale",
  placement = "right",
  size = "xs"
)
}
\arguments{
\item{size}{}
}
\value{

}
\description{
Title
}
