% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraExample.R
\name{chakraExample}
\alias{chakraExample}
\title{Run a Chakra example}
\usage{
chakraExample(example, display.mode = "showcase", ...)
}
\arguments{
\item{example}{example name}

\item{display.mode}{the display mode to use when running the example; see
\code{\link[shiny:runApp]{runApp}}}

\item{...}{arguments passed to \code{\link[shiny:runApp]{runApp}}}
}
\description{
A function to run examples of Shiny apps with Chakra components.
}
\examples{
if(interactive()){
  chakraExample("Menu")
}
}
