% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraExample.R
\name{chakraExamples}
\alias{chakraExamples}
\title{Chakra examples}
\usage{
chakraExamples()
}
\value{
No return value, only prints a message listing the Chakra examples.
}
\description{
List of Chakra examples.
}
\examples{
chakraExamples()
if(interactive()){
  chakraExample("MenuWithGroups")
}
}
