% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraIcons}
\alias{chakraIcons}
\title{Chakra icons}
\usage{
chakraIcons()
}
\description{
List of chakra icons.
}
\details{
See \href{https://chakra-ui.com/docs/media-and-icons/icon#all-icons}{all chakra icons}.
}
\examples{
chakraIcons()
}
