% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraModal.R
\name{chakraModalOptions}
\alias{chakraModalOptions}
\title{Modal options}
\usage{
chakraModalOptions(
  closeOnEsc = TRUE,
  isCentered = TRUE,
  motionPreset = "scale",
  size = "md",
  ...
)
}
\arguments{
\item{closeOnEsc}{whether to close the modal on pressing the 'esc' key}

\item{isCentered}{whether to center the modal on screen}

\item{motionPreset}{transition that should be used for the modal; one of
\code{"scale"}, \code{"none"}, \code{"slideInBottom"}, or
\code{"slideInRight"}}

\item{size}{modal size, \code{"sm"}, \code{"md"}, \code{"lg"}, \code{"xl"},
\code{"2xl"}, \code{"full"}, \code{"xs"}, \code{"3xl"}, \code{"4xl"},
\code{"5xl"}, or \code{"6xl"}}

\item{...}{other attributes of \code{Modal}}
}
\value{
A named list, for usage in \code{\link{chakraModal}}.
}
\description{
Options for the modal widget (\code{\link{chakraModal}}).
}
