% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jseval_and_hooks.R
\name{getHookProperty}
\alias{getHookProperty}
\title{Get hook property}
\usage{
getHookProperty(hook, property)
}
\arguments{
\item{hook}{the name of the hook, usually created in the \code{states} list
of the \code{\link{withStates}} function}

\item{property}{the hook property you want to get}
}
\description{
Chakra hooks are JavaScript objects; this function allows to
  get a property (key) of a hook. See \code{\link{useDisclosure}} for an
  example.
}
