% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Icons.R
\name{ionIcons}
\alias{ionIcons}
\title{Ionicons}
\usage{
ionIcons()
}
\value{
The names of the ionicons in a vector.
}
\description{
List of ionicons.
}
\details{
See \href{https://ionic.io/ionicons}{ionicons website}.
}
\examples{
ionIcons()
}
