% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraCombinedSlider.R
\name{numberInputOptions}
\alias{numberInputOptions}
\title{Options for the number input of the combined Chakra slider}
\usage{
numberInputOptions(
  precision = NULL,
  maxWidth = "80px",
  fontSize = NULL,
  fontColor = NULL,
  borderColor = NULL,
  focusBorderColor = NULL,
  borderWidth = NULL,
  incrementStepperColor = NULL,
  decrementStepperColor = NULL,
  ...
)
}
\arguments{
\item{precision}{number of decimal points}

\item{maxWidth}{width of the number input, e.g. \code{"100px"} or
\code{"20\%"}}

\item{fontSize}{font size of the displayed value, e.g. \code{"15px"}}

\item{fontColor}{color of the displayed value}

\item{borderColor}{color of the border of the number input}

\item{focusBorderColor}{color of the border of the number input on focus}

\item{borderWidth}{width of the border of the number input,
e.g. \code{"3px"} or \code{"medium"}}

\item{incrementStepperColor}{color of the increment stepper}

\item{decrementStepperColor}{color of the decrement stepper}

\item{...}{other attributes of \code{NumberInput}}
}
\description{
Create a list of options to be passed to the
  \code{numericInputOptions} argument in \code{\link{chakraCombinedSlider}}.
}
