% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{import}
\name{shinyChakraUI-imports}
\alias{shinyChakraUI-imports}
\alias{React}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages.
  Follow the links to their documentation:
  \code{\link[reactR:React]{React}}.
}
