% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraSlider.R
\name{sliderMarkOptions}
\alias{sliderMarkOptions}
\title{Slider mark options.}
\usage{
sliderMarkOptions(
  textAlign = "center",
  backgroundColor = "blue.500",
  textColor = "white",
  margin = "-35px 0 0 -25px",
  padding = "0 10px",
  width = "50px",
  ...
)
}
\arguments{
\item{textAlign}{text alignment}

\item{backgroundColor}{background color}

\item{textColor}{text color}

\item{margin}{margin (CSS property)}

\item{padding}{padding (CSS property)}

\item{width}{width}

\item{...}{other attributes}
}
\value{
A list of attributes for usage in \code{\link{chakraSlider}}.
}
\description{
Define the options for the slider mark.
}
