% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraSlider.R
\name{sliderThumbOptions}
\alias{sliderThumbOptions}
\title{Slider thumb options}
\usage{
sliderThumbOptions(
  width = NULL,
  height = NULL,
  color = NULL,
  borderColor = NULL,
  borderWidth = NULL
)
}
\arguments{
\item{width}{thumb width}

\item{height}{thumb height}

\item{color}{thumb color}

\item{borderColor}{thumb border color}

\item{borderWidth}{thumb border width}
}
\value{
A list of attributes for usage in \code{\link{chakraSlider}}.
}
\description{
Define the slider thumb options.
}
