% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraSlider.R
\name{sliderTooltipOptions}
\alias{sliderTooltipOptions}
\title{Title}
\usage{
sliderTooltipOptions(
  hasArrow = TRUE,
  backgroundColor = "red.600",
  placement = "top",
  closeOnClick = FALSE,
  isOpen = TRUE,
  ...
)
}
\arguments{
\item{hasArrow}{whether to include an arrow}

\item{backgroundColor}{background color}

\item{placement}{tooltip placement; see
\href{https://chakra-ui.com/docs/overlay/tooltip#placement}{tooltip placement}}

\item{closeOnClick}{whether to close the tooltip on click}

\item{isOpen}{whether the tooltip is open}

\item{...}{other attributes}
}
\value{
A list of attributes for usage in \code{\link{chakraSlider}}.
}
\description{
Title
}
