% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Icons.R
\name{superTinyIcons}
\alias{superTinyIcons}
\title{Super tiny icons}
\usage{
superTinyIcons()
}
\value{
The names of the super tiny icons in a vector.
}
\description{
List of super tiny icons.
}
\details{
See \href{https://github.com/edent/SuperTinyIcons#whats-available-so-far}{all super tiny icons}.
}
\examples{
superTinyIcons()
}
