% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tag.R
\docType{data}
\name{Tag}
\alias{Tag}
\title{React component builder}
\format{
An object of class \code{ReactTagBuilder} of length 0.
}
\usage{
Tag
}
\description{
Create a React component. This is similar to
  \code{\link[reactR:React]{React}}.
}
\examples{
Tag$Box(
  bg = "tomato",
  Tag$ButtonGroup(
    spacing = "4",
    Tag$Button(
      "I'm a button"
    ),
    Tag$Button(
      "I'm another button"
    )
  )
)
}
\keyword{datasets}
