% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jseval_and_hooks.R
\name{useRef}
\alias{useRef}
\title{The `useRef` hook}
\usage{
useRef(initialValue = NA)
}
\arguments{
\item{initialValue}{optional initial value}
}
\value{
A list like the return value of \code{\link{jseval}}.
}
\description{
The React `useRef` hook.
}
