% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{withStates}
\alias{withStates}
\title{Component with states or hooks}
\usage{
withStates(component, states)
}
\arguments{
\item{component}{a React component}

\item{states}{list of states}
}
\value{
A component to use in \code{\link{chakraComponent}}.
}
\description{
Create a component with React states and/or hooks.
}
\examples{
library(shiny)
library(shinyChakraUI)

ui <- chakraPage(

  br(),

  chakraComponent(
    "mycomponent",

    withStates(

      Tag$Fragment(

        Tag$Container(
          maxW = "xl",
          centerContent = TRUE,
          bg = "orange.50",
          Tag$Heading(
            getState("heading")
          ),
          Tag$Text(
            "I'm just some text."
          )
        ),

        br(),
        Tag$Divider(),
        br(),

        Tag$Button(
          colorScheme = "twitter",
          display = "block",
          onClick = jseval(
            "() => setState('heading', 'I am the new heading.')"
          ),
          "Click me to change the heading"
        )

      ),

      states = list(heading = "I am the heading.")

    )

  )

)

server <- function(input, output, session){}

if(interactive()){
  shinyApp(ui, server)
}
}
