% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraAlertDialog}
\alias{chakraAlertDialog}
\title{Title}
\usage{
chakraAlertDialog(
  inputId,
  options = chakraAlertDialogOptions(),
  openButton,
  header,
  body,
  footerButtons = Tag$Button("Cancel", action = "cancel", value = "cancel")
)
}
\arguments{
\item{inputId}{}
}
\value{

}
\description{
Title
}
