% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraAlertDialogOptions}
\alias{chakraAlertDialogOptions}
\title{Title}
\usage{
chakraAlertDialogOptions(
  closeOnEsc = TRUE,
  colorScheme = "red",
  isCentered = TRUE,
  motionPreset = "scale",
  size = "md"
)
}
\arguments{
\item{size}{}
}
\value{

}
\description{
Title
}
