% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraComponent}
\alias{chakraComponent}
\title{Chakra component}
\usage{
chakraComponent(componentId, ...)
}
\arguments{
\item{componentId}{component id}

\item{...}{elements to include within the component}
}
\description{
Create a Chakra component.
}
