% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Icons.R
\name{chakraIcons}
\alias{chakraIcons}
\title{Chakra icons}
\usage{
chakraIcons()
}
\value{
The names of the Chakra icons in a vector.
}
\description{
List of Chakra icons.
}
\details{
See \href{https://chakra-ui.com/docs/media-and-icons/icon#all-icons}{all chakra icons}.
}
\examples{
chakraIcons()
}
