% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{chakraPage}
\alias{chakraPage}
\title{Chakra page}
\usage{
chakraPage(...)
}
\arguments{
\item{...}{elements to include within the page}
}
\description{
Function to be used as the \code{ui} element of a Shiny app;
  it is intended to contain some \code{\link{chakraComponent}} elements.
}
