% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jseval_and_hooks.R
\name{jseval}
\alias{jseval}
\title{Evaluate JS code}
\usage{
jseval(code)
}
\arguments{
\item{code}{JavaScript code given as a string}
}
\value{
A list containing the URL-encoded JavaScript code.
}
\description{
Evaluate JavaScript code in the application.
}
\examples{
library(shiny)
library(shinyChakraUI)

ui <- chakraPage(

  br(),

  chakraComponent(
    "mycomponent",

    Tag$Button(
      colorScheme = "pink",
      size = "lg",
      onClick = jseval('() => alert("Hello Chakra")'),
      "Trigger alert"
    )

  )

)

server <- function(input, output, session){}

if(interactive()){
  shinyApp(ui, server)
}
}
