% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakra.R
\name{withStates}
\alias{withStates}
\title{Component with states or hooks}
\usage{
withStates(component, states)
}
\arguments{
\item{component}{a React component}

\item{states}{list of states}
}
\value{
A component to use in \code{\link{chakraComponent}}.
}
\description{
Create a component with React states and/or hooks.
}
